Create a binary classifier for Toxic dataset
Using a simple Naive Bayes model, an initial classifier can be trained,
verified, and used to test standalone comments belonging to a single
value in a binary domain: toxic, or non-toxic. This model is
lightweight and simple, and is intended to determine whether a comment
is toxic or not as quickly as possible, with the intention that toxic
comments can be flagged in near-realtime.

Signed-Off-By: Robert Clark <robdclark@outlook.com>